add super class in condition view
