+++ app/frontend/javascripts/definition.js
@@ -1,0 +1,5 @@
+
+export const conditionItemType = {
+  condition: 0,
+  group: 1
+}
