+++ app/frontend/stylesheets/foundation/_base.sass
@@ -78,6 +78,9 @@ $Z_INDEX_CONFIGURATION: 4
 $Z_INDEX_EDITING_CONDITION: 4
 $Z_INDEX_ACTIVITY_INDICATOR: 5
 
+:root
+  
+
 $BACKDROP_BG: radial-gradient(ellipse at center, rgba($COLOR_KEY, .3) 10%, rgba($COLOR_KEY, .2) 60%, rgba($COLOR_KEY_DARK3, .1) 120%)
 
 // time
