+++ app/frontend/stylesheets/object/component/_advanced-search-condition-view.sass
@@ -98,14 +98,6 @@ $CORNER_RADIUS: $HEIGHT / 2
     opacity: 0
     pointer-events: none
     transition: opacity $TRANSITION_DURATION
-  &.-selected
-    > .body
-      > .summary
-        background-color: $COLOR_SELECTED_LIGHT
-        > .classification
-          background-color: $COLOR_SELECTED
-          color: $COLOR_BLACK
-
   &.-editing
     > .body
       z-index: $Z_INDEX_EDITING_CONDITION + 1
@@ -156,6 +148,13 @@ $CORNER_RADIUS: $HEIGHT / 2
             transform: rotate(-20deg)
             top: 2px
             left: 6.5px
+  &.-selected
+    > .body
+      > .summary
+        background-color: $COLOR_SELECTED_LIGHT
+        > .classification
+          background-color: $COLOR_SELECTED
+          color: $COLOR_BLACK
   &[data-classification="frequency"], &[data-classification="count"]
     > .body
       > .summary
