+++ app/frontend/stylesheets/object/component/_advanced-search-toolbar.sass
@@ -1,7 +1,7 @@
 $OPACITY_DISABLED: .5
 
 .advanced-search-toolbar
-  margin: 4px 0 -1px -14px
+  margin: 4px 0 0 calc(var(--height-toolbar) * -1)
   padding: 2px 6px
   background-color: $COLOR_KEY_DARK3
   color: white
