+++ app/frontend/stylesheets/object/project/_AdvancedSearchBuilderView.sass
@@ -2,3 +2,6 @@
   padding: 14px $PADDING_SIDE 14px
   // > .inner
   //   > .buildersection
+  --padding-group: 8px
+  --height-toolbar: 14px
+
